Changed dto range to real numbers

+++ CSharpDB/02.EntityFrameworkCore/Exams/C#DBAdvancedRetakeExam-11April2023/Invoices/DataProcessor/DTOs/Import/ProductImportDTO.cs
@@ -14,7 +14,7 @@ namespace Invoices.DataProcessor.DTOs.Import
         public string Name { get; set; } = null!;
 
         [Required]
-        [Range(ProductPriceMinLength, ProductPriceMaxLength)]
+        [Range((double)ProductPriceMinLength, (double)ProductPriceMaxLength)]
         [JsonProperty(nameof(Price))]
         public decimal Price { get; set; }
